+++ packages/vda-did/CHANGELOG.md
@@ -1,3 +1,8 @@
+
+
+- Updated to use `@verida/vda-common`
+- Update tests
+
 2023-03-27 (v2.2.1)
 -------------------